fixup(lazy)!: Get rid of renamig most of plugins.
This would cause some plugins to be unable to search
specific plugins... Therefore, we'll have to stick with the
repo names for most of them. This behaviour will be
checked further.

Make sure to Sync, Clean and Build your plugins. Also, if you're
using `neorg`, run `:Neorg sync-parsers` if using TS.

Signed-off-by: Guennadi Maximov C <g.maxc.fox@protonmail.com>